Remove cast to int from call to getEntropyInput

+++ DRBG.php
@@ -35,7 +35,7 @@ abstract class DRBG {
     private function instantiate() {
         
         try {    
-            $entropy = self::__getEntropyInput((int) 1.5 * self::STRENGTH);
+            $entropy = self::__getEntropyInput(1.5 * self::STRENGTH);
             $nonce = self::__getEntropyInput(self::STRENGTH / 2);
             $this->instantiateAlgorithm($entropy, $nonce);
         } catch (Exception $e) {
